requirements.txt
Flask==2.3.3
google-generativeai>=0.4.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
gunicorn==21.2.0
pytest==7.4.3
flake8==6.1.0
pytest-mock==3.12.0
